src/requirements.txt
datasets
nltk
sentencepiece
transformers
numpy
rouge
torchmetrics